==> dataset_generator_pmd/1_PMD_dbnsfp.ipynb <==
import os
import sys
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
home_dir = "../"

import pandas as pd
from typing import List

def summarize_functional_effect(mut_function:str):
    # print(mut_function)
    effects = []
    effect = ""
    flag = False
    for i, ch in enumerate(mut_function):
        if ch=='[':
            flag = True
            continue
        if flag and ch in ['-', '+'] and mut_function[i+1] in ['-', '+', ']', ' ']:
            effect += ch
        elif flag and ch == '=':
            effects.append('=')
        elif flag and ch == '0' and mut_function[i+1]=="]":
            effects.append('0')
            
        if ch=="]":
            flag = False
            if len(effect)>0: effects.append(effect)
            effect=""
    
    if len(effects)>0: return effects
    else: return ""

def get_functional_effect(effects:List):
    counts = [] 
    functional_effect = ""
    for ef in effects:
        if "0" in ef:
            return "Knock-out"
        elif "=" in ef:
            functional_effect = "No-effect"
        else:
            counts.append(len(ef)-2)

    
    if len(counts)==0 and functional_effect=="No-effect": return "No-effect"
    elif max(counts)<3: return "Effect"
    else: return "Knock-out"

data_path = home_dir+f"data/pmd/dfPMD-NSFP43a_subset_func.tsv.gz"
pmd_df = pd.read_csv(data_path, sep="\t", compression="gzip")
print(pmd_df.shape)
print(pmd_df.columns.to_list())

pmd_df["wt"] = pmd_df["mut_real"].apply(lambda x: x[0]) # 1-letter amino acid
pmd_df["mut"] = pmd_df["mut_real"].apply(lambda x: x[-1])
pmd_df["prot_pos"] = pmd_df["mut_real"].apply(lambda x: int(x[1:-1])) # mutation position is 1-indexed

pmd_df["function_summarized"] = pmd_df["function"].apply(summarize_functional_effect)
pmd_df = pmd_df[pmd_df["function_summarized"].apply(lambda x:  len(x)>0)] # removing rows that does not have any functional effect annotations
pmd_df["functional_effect"] = pmd_df["function_summarized"].apply(get_functional_effect)
print("after removing non-functional annotations:", pmd_df.shape)

pmd_df["pmd_nr_id"] = pmd_df["pmd_id"]+pmd_df["nr"].apply(lambda x: "_"+str(int(x)))

pmd_df = pmd_df[pmd_df["source"].apply(lambda x: True if "Human" in str(x) else False)] # only keeping human variants
print("after removing non-human entries:", pmd_df.shape)

# pmd_df.head()
print(pmd_df["functional_effect"].value_counts())
print("#-of unique proteins", pmd_df[["pmd_id", "nr"]].drop_duplicates(keep="first").shape) # 2263
print("#-of unique genes", pmd_df["genename"].drop_duplicates(keep="first").shape) # 1089
print(pmd_df.shape)

info_cols = ['mut_id', 'md5', 'pmd_id', 'nr', 'pmd_nr_id', 'crossref', 'uniprot_id', 'protein', 'source', 'mut_PMD', 'mut_real', 'wt', 'mut', 'prot_pos', '#chr', 'pos(1-based)', 'ref', 'alt', 'rs_dbSNP', 'genename', 'ensembl_id', 'function', 'function_summarized', 'functional_effect']
method_score_cols = ["CADD_raw", "MetaRNN_score", "MVP_score", "Polyphen2_HVAR_score", "REVEL_score", "SIFT_score", "integrated_fitCons_score", "phyloP17way_primate", "phastCons17way_primate", "bStatistic"]
other_cols = ['seq']
columns = info_cols + method_score_cols + other_cols
pmd_df = pmd_df[columns]

# filtering the dataset on the protein sequence len <=1022
sequences_df = pmd_df[['pmd_nr_id', 'crossref', 'seq']].drop_duplicates(keep="first")
print("#-of sequences b4 seq-len filter", sequences_df.shape[0])
protid_seq_tuple_list = [(tuple.pmd_nr_id, tuple.seq) for tuple in sequences_df.itertuples() if len(str(tuple.seq))<=1022]
print("#-of sequences after seq-len filter", len(protid_seq_tuple_list))

new_protein_acc_list = list(zip(*protid_seq_tuple_list))[0]
pmd_df = pmd_df[pmd_df["pmd_nr_id"].isin(new_protein_acc_list)]
print("after filtering on seq-len <=1022:", pmd_df.shape)
print(pmd_df["functional_effect"].value_counts())

print("#-of unique proteins", pmd_df[["pmd_id", "nr"]].drop_duplicates(keep="first").shape) # 2068
print("#-of unique genes", pmd_df["genename"].drop_duplicates(keep="first").shape) # 997

# computing missing rows for each methods
import numpy as np    
def compute_avg(x):
    x = str(x).split(";")
    return np.mean([float(i) for i in x if i!="."]) 

def compute_number_of_missing_rows(col_name): 
    avg_scores = pmd_df[col_name].apply(compute_avg)
    avg_scores = avg_scores[pd.isna(avg_scores)]
    n_missing_rows = avg_scores.shape[0]
    total_rows = pmd_df.shape[0]
    print(col_name, f"\t100*({n_missing_rows}/{total_rows})=", f"{100*(n_missing_rows/total_rows):.4f}")

for col_name in method_score_cols:
    compute_number_of_missing_rows(col_name)
    # break

# pmd_df["MVP_score"].value_counts()

pmd_df.to_csv(home_dir+"data/pmd/pmd_with_dbnsfp.tsv", sep="\t", index=False)

### testing different things here...

pmd_df[["pmd_nr_id", "prot_pos"]].value_counts()
pmd_df[(pmd_df["pmd_nr_id"]=="A000006_2") & (pmd_df["prot_pos"]==271)][["wt", "mut", "prot_pos"]]

print("#-of rows mapped to rs-ids", pmd_df[pmd_df["rs_dbSNP"]!="."].shape)  
print("#-rows not mapped to rs-ids", pmd_df[pmd_df["rs_dbSNP"]=="."].shape)

print(pmd_df[pmd_df["crossref"].apply(lambda x: False if "HUMAN" in x else True)]["crossref"].unique())
print(pmd_df[pmd_df["source"].apply(lambda x: False if "Human" in str(x) else True)]["source"].unique())
pmd_df[pmd_df["source"].apply(lambda x: True if "Human" in str(x) else False)].shape
