==> dataset_generator_pmd/2_PMD_merging_with_rsids.ipynb <==
import os
import sys
home_dir = "../"
module_path = os.path.abspath(os.path.join(home_dir))
if module_path not in sys.path:
    sys.path.append(module_path)

import pandas as pd
from typing import List
# see notes/mysql to find from where and how the data [SNPs_dbSNP, SNPs_PMD and seqs_PMD] is downloaded

# extract functional annotation
# def summarize_functional_effect(mut_function):
#     # print(mut_function)
#     mut_function = str(mut_function)
#     effects = []
#     effect = ""
#     flag = False
#     for i, ch in enumerate(mut_function):
#         if ch=='[':
#             flag = True
#             continue
#         if flag and ch in ['-', '+'] and mut_function[i+1] in ['-', '+', ']', ' ']:
#             effect += ch
#         elif flag and ch == '=':
#             effects.append('=')
#         elif flag and ch == '0' and mut_function[i+1]=="]":
#             effects.append('0')
            
#         if ch=="]":
#             flag = False
#             if len(effect)>0: effects.append(effect)
#             effect=""
    
#     if len(effects)>0: return effects
#     else: return ""

# def get_functional_effect(effects):
#     counts = [] 
#     functional_effect = ""
#     for ef in effects:
#         if "0" in ef:
#             return "Knock-out"
#         elif "=" in ef:
#             functional_effect = "No-effect"
#         else:
#             counts.append(len(ef)-2)

    
#     if len(counts)==0 and functional_effect=="No-effect": return "No-effect"
#     elif max(counts)<3: return "Effect"
#     else: return "Knock-out"

# loading SNPs_PMD table
raw_pmd_df = pd.read_csv(home_dir+"data/SNPdbe/SNPs_PMD.csv", sep="\t;") # PMD: from SNPdbe # (127565, 48)
raw_pmd_df = raw_pmd_df[(raw_pmd_df["function"]!=";\\N") & (raw_pmd_df["function"]!="\\N")]  # (72738, 48)
raw_pmd_df = raw_pmd_df[raw_pmd_df["function"].apply(lambda x: "[" in x)] # removing rows that does not have functional effect annotations (65128, 48)
# print(raw_pmd_df.shape) 

# raw_pmd_df = raw_pmd_df.copy(deep=True)
# raw_pmd_df["wt"] = raw_pmd_df["mut_real"].apply(lambda x: x[0]) # 1-letter amino acid
# raw_pmd_df["mut"] = raw_pmd_df["mut_real"].apply(lambda x: x[-1])
# raw_pmd_df["prot_pos"] = raw_pmd_df["mut_real"].apply(lambda x: int(x[1:-1])) # mutation position is 1-indexed

# raw_pmd_df["function_summarized"] = raw_pmd_df["function"].apply(summarize_functional_effect)
# raw_pmd_df = raw_pmd_df[raw_pmd_df["function_summarized"].apply(lambda x:  len(x)>0)] # removing rows that does not have any functional effect annotations (64750, 50)
# raw_pmd_df["functional_effect"] = raw_pmd_df["function_summarized"].apply(get_functional_effect)

print(raw_pmd_df.shape) 
print(raw_pmd_df.columns)

### PMD data merging with gene information
gene_df = pd.read_csv(home_dir+"data/SNPdbe/md52keywords.tsv", sep="\t")
print(gene_df.shape) 
print(gene_df.columns)
gene_df = gene_df[(gene_df["gene_SP"]!=";\\N") & (gene_df["gene_SP"]!="\\N") & (~pd.isna(gene_df["gene_SP"]))] # (147349, 4)
print(gene_df.shape) 
print(gene_df.columns)
pmd_df = pd.merge(left=raw_pmd_df, right=gene_df, on="md5", how="left")
print(pmd_df.shape) 
print(pmd_df.columns)

print(pmd_df["gene_SP"].unique().shape)
pmd_df["gene_SP"].value_counts()

### PMD data merging with corresponding sequences
# seq_pmd_df = pd.read_csv(home_dir+"data/SNPdbe/seqs_PMD.csv", sep="\t;") # PMD: from SNPdbe # (127565, 48)
# seq_pmd_df.drop(columns=["taxid", "md5"], inplace=True)
# print(seq_pmd_df.shape) 
# print(seq_pmd_df.columns)

# pmd_df = pd.merge(left=pmd_df, right=seq_pmd_df, on=["pmd_id", "nr"], how="inner")
# # pmd_columns = ["mut_id", "pmd_id", "nr", "crossref", "uniprot_id", "ensembl_id", "taxid", "protein", "mut_PMD", "mut_real", 'wt', 'mut', 'prot_pos', 'function_summarized', 'functional_effect', "function", "seq"]
# # pmd_df = pmd_df[pmd_columns]
# print(pmd_df.shape) 
# print(pmd_df.columns)
# pmd_df[["pmd_id", "mut_real", "seq"]]

# # seq = seq_pmd_df[(seq_pmd_df["pmd_id"]=="A000006") & (seq_pmd_df["nr"]==2)]["seq"].values[0]
# # print(seq[270])
# # pmd_df.iloc[0]

# loading PMD SNPs_dbSNP table
dbSNP_df = pd.read_csv(home_dir+"data/SNPdbe/SNPs_dbSNP.csv", sep="\t;") # dbSNP: from SNPdbe
dbSNP_df = dbSNP_df[(dbSNP_df["mut_id"]!=";\\N") & (dbSNP_df["mut_id"]!="\\N")] # (1780330, 34)
dbSNP_df.drop(columns=["taxid"], inplace=True)

dbSNP_df = dbSNP_df.astype({'mut_id': 'int64'})
print(dbSNP_df.shape) # (1526352, 34)
# print(dbSNP_df.columns)

# column description can be found here: https://www.rostlab.org/services/snpdbe/docu/schema.pdf
# protein pos is 0-indexed
dbSNP_cols = ["mut_id", "snp_id", "mrna_acc", "mrna_ver", "mrna_pos", "allele",  "protein_acc", "protein_ver", "verified"]
dbSNP_df = dbSNP_df[dbSNP_cols]
print(dbSNP_df.columns)

mutid_snpid_pair_df = dbSNP_df[["mut_id", "snp_id"]].drop_duplicates(keep="first")
mutid_snpid_pair_df.shape

# mapping PMD rows to rs-ids
pmd_with_rsids_df = pd.merge(left=pmd_df, right=mutid_snpid_pair_df, how="inner", on="mut_id")
print(pmd_with_rsids_df.shape)
print(pmd_with_rsids_df.columns)

### Download SNPs using rs-id list (work on this only once)
snp_ids = pmd_with_rsids_df[~pd.isna(pmd_with_rsids_df["snp_id"])]["snp_id"].unique()
print(len(snp_ids))
snp_ids = [str(int(snp_id)) for snp_id in snp_ids]
snp_ids = " ".join(snp_ids)
snp_ids # 1. copy output and run download_snps_from_snplist (did not work) 2. manually downloaded and merged 100 rs-ids at a time.

### Loading downloaded SNPs using the above rs-ids
snps_df = pd.read_csv(home_dir+"data/SNPdbe/snps.tsv", sep="\t", comment="#")
print(snps_df.columns)
snps_df = snps_df[['chr', 'pos', 'variation', 'variant_type', 'snp_id']]
print(snps_df.columns)
# print(pmd_snps_df.head())

variations = []
for i, tuple in enumerate(snps_df.itertuples()):
    if tuple.variant_type != "snv": # only considering SNVs
        print(tuple)
        # break
        continue
    
    ref_allele, alt_allele = tuple.variation.split(">")
    alt_alleles = alt_allele.split(",")
    # print(ref_allele, ">>>", alt_alleles)

    for alt_allele in alt_alleles:
        row = dict(snps_df.loc[i])
        row["ref_allele"] = ref_allele
        row["alt_allele"] = alt_allele
        variations.append(row)

snps_df = pd.DataFrame(variations)
snps_df = snps_df[~pd.isna(snps_df["chr"])]
snps_df = snps_df[~pd.isna(snps_df["pos"])]
snps_df = snps_df[~pd.isna(snps_df["snp_id"])]
snps_df.rename(columns={"pos": "chrom_pos", "chr": "chrom"}, inplace=True)
snps_df

# merging chromosomal information from snp_id 
pmd_data_df = pd.merge(left=pmd_with_rsids_df, right=snps_df, on="snp_id", how="inner") # multiple rows incurred for SNPs data under same snp_id
pmd_data_df.drop_duplicates(keep="first", inplace=True, ignore_index=True)
print(pmd_data_df.shape)
print(pmd_data_df.columns)
pmd_data_df.to_csv(home_dir+"data/pmd/pmd_with_rsids.tsv", sep="\t", index=False)
# pmd_data_df.value_counts()

# pmd_data_df.to_csv(home_dir+"models/aa_common/datasets_pmd_analysis/pmd_data.tsv", sep="\t", index=False)
# sequences_df = pmd_data_df[['pmd_nr_id', 'crossref', 'seq']].drop_duplicates(keep="first")
# print("#-of sequences", sequences_df.shape)

# out_fasta_filepath = home_dir+"models/aa_common/datasets_pmd_analysis/pmd_sequences.fasta"
# out_fasta_file_handle = open(out_fasta_filepath, "w")

# for tuple in sequences_df.itertuples():
#     out_fasta_file_handle.write(f">{tuple.pmd_nr_id} | {tuple.crossref}\n")
#     out_fasta_file_handle.write(f"{tuple.seq}\n")
#     # break
# out_fasta_file_handle.close()  

# print(pmd_data_df["functional_effect"].value_counts())
# print(pmd_data_df[pd.isna(pmd_data_df["mut_real"])].shape)
# print(pmd_data_df[pd.isna(pmd_data_df["function"])].shape)
# print(pmd_data_df[pd.isna(pmd_data_df["functional_effect"])].shape)
# print("#-snps", pmd_data_df[~pd.isna(pmd_data_df["snp_id"])].shape) # number of rows that is mapped to snp-id: 2877
# print("#-human (crossref)", pmd_data_df[pmd_data_df["crossref"].apply(lambda x: True if "HUMAN" in x else False)].shape) # number of human entries: 20594
# print("#-human (taxid)", pmd_data_df[pmd_data_df["taxid"]=="9606"].shape) # # number of human entries using taxid: 16088
# pmd_data_df[pmd_data_df["crossref"].apply(lambda x: True if "HUMAN" in x else False)][['pmd_id', 'nr']].drop_duplicates(keep="first")
# # print(merged_df[(merged_df["crossref"].apply(lambda x: True if "HUMAN" in x else False)) & ~pd.isna(merged_df["snp_id"])].shape) # number of human entries that is mapped to rs-id: 2829

# print("#-of unique genes: ", pmd_data_df["gene_SP"].unique().shape)
