==> dataset_generator_pmd/3_PMD_merging_dbnsfp_and_rsids.ipynb <==
import os
import sys
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
home_dir = "../"

import pandas as pd

pmd_with_rsids_df = pd.read_csv(home_dir+"data/pmd/pmd_with_rsids.tsv", sep="\t")
pmd_with_rsids_df["snp_id"] = pmd_with_rsids_df["snp_id"].apply(lambda x: "rs"+str(int(x)))
pmd_with_rsids_df.rename(columns={'gene_SP':'genename'}, inplace=True)
pmd_with_rsids_df["genename"] = pmd_with_rsids_df["genename"].apply(lambda x: str(x).split("_")[0])
print(pmd_with_rsids_df.shape)
print(pmd_with_rsids_df.columns)

# pmd_with_rsids_df["genename"]

pmd_dbnsfp_df = pd.read_csv(home_dir+"data/pmd/pmd_with_dbnsfp.tsv", sep="\t")
pmd_dbnsfp_df.rename(columns={'#chr':'chrom', 'pos(1-based)':'chrom_pos', 'ref':'ref_allele', 'alt':'alt_allele', 'rs_dbSNP':'snp_id'}, inplace=True)
pmd_dbnsfp_df.drop_duplicates(keep="first", inplace=True, ignore_index=True)
pmd_dbnsfp_df["snp_id"].replace(".", None, inplace=True)
pmd_dbnsfp_df["genename"] = pmd_dbnsfp_df["genename"].apply(lambda x: x.split(";")[0])
print(pmd_dbnsfp_df.shape)
print(pmd_dbnsfp_df.columns)

# pmd_dbnsfp_df["genename"]

columns = ['mut_id', 'md5', 'pmd_id', 'nr', 'mut_real', 'chrom', 'chrom_pos', 'ref_allele', 'alt_allele', 'function', 'source', 'crossref', 'snp_id']
x = pmd_with_rsids_df[columns]
y = pmd_dbnsfp_df[columns]
z = pd.merge(x, y, how='outer', on=columns)
z.drop_duplicates(keep="first", inplace=True, ignore_index=True)
z.value_counts()

for col in columns:
    print(col, z[pd.isna(z[col])].shape)

pmd_clean_df = z.copy(deep=True)
print(pmd_clean_df.shape)

# merging with raw PMD-with-rsids data
data_merged_with_pmd_raw = pd.merge(left=pmd_clean_df, right=pmd_with_rsids_df, how="left", on=columns)
print(data_merged_with_pmd_raw.shape)
print(data_merged_with_pmd_raw.columns)

# merging with PMD dbnsfp data
data_merged_with_pmd_dbnsfp = pd.merge(left=data_merged_with_pmd_raw, right=pmd_dbnsfp_df, how="left", on=columns)
print(data_merged_with_pmd_dbnsfp.shape)
print(data_merged_with_pmd_dbnsfp.columns)

method_score_cols = ["CADD_raw", "MetaRNN_score", "MVP_score", "Polyphen2_HVAR_score", "REVEL_score", "SIFT_score", "integrated_fitCons_score", "phyloP17way_primate", "phastCons17way_primate", "bStatistic"]
x_columns = columns + method_score_cols
data_merged_with_pmd_dbnsfp = data_merged_with_pmd_dbnsfp[x_columns]
print(data_merged_with_pmd_dbnsfp.shape)
print(data_merged_with_pmd_dbnsfp.columns)

# loading PMD seq table
seq_pmd_df = pd.read_csv(home_dir+"data/SNPdbe/seqs_PMD.csv", sep="\t;") # PMD: from SNPdbe # (127565, 48)
seq_pmd_df.drop(columns=["taxid", "md5"], inplace=True)
print(seq_pmd_df.shape) 
print(seq_pmd_df.columns)

# merging with seq information
data_merged_with_seq = pd.merge(left=data_merged_with_pmd_dbnsfp, right=seq_pmd_df, on=["pmd_id", "nr"], how="inner")
print(data_merged_with_seq.shape) 
print(data_merged_with_seq.columns)

# extract functional annotation
def summarize_functional_effect(mut_function):
    # print(mut_function)
    mut_function = str(mut_function)
    effects = []
    effect = ""
    flag = False
    for i, ch in enumerate(mut_function):
        if ch=='[':
            flag = True
            continue
        if flag and ch in ['-', '+'] and mut_function[i+1] in ['-', '+', ']', ' ']:
            effect += ch
        elif flag and ch == '=':
            effects.append('=')
        elif flag and ch == '0' and mut_function[i+1]=="]":
            effects.append('0')
            
        if ch=="]":
            flag = False
            if len(effect)>0: effects.append(effect)
            effect=""
    
    if len(effects)>0: return effects
    else: return ""

def get_functional_effect(effects):
    counts = [] 
    functional_effect = ""
    for ef in effects:
        if "0" in ef:
            return "Knock-out"
        elif "=" in ef:
            functional_effect = "No-effect"
        else:
            counts.append(len(ef)-2)

    
    if len(counts)==0 and functional_effect=="No-effect": return "No-effect"
    elif max(counts)<3: return "Effect"
    else: return "Knock-out"

df = data_merged_with_seq.copy(deep=True)
df["wt"] = df["mut_real"].apply(lambda x: x[0]) # 1-letter amino acid
df["mut"] = df["mut_real"].apply(lambda x: x[-1])
df["prot_pos"] = df["mut_real"].apply(lambda x: int(x[1:-1])) # mutation position is 1-indexed

df["function_summarized"] = df["function"].apply(summarize_functional_effect)
df = df[df["function_summarized"].apply(lambda x:  len(x)>0)] # removing rows that does not have any functional effect annotations
df["functional_effect"] = df["function_summarized"].apply(get_functional_effect)

data_filered_on_func = df.copy(deep=True)
print(data_filered_on_func.shape) 
print(data_filered_on_func.columns)

# filter on human
df = data_filered_on_func.copy(deep=True)
df = df[df["source"].apply(lambda x: True if "Human" in str(x) else False)] # only keeping human variants
print("after removing non-human entries:", df.shape)

data_filered_on_human = df.copy(deep=True)

# filter on seq-len
df = data_filered_on_human.copy(deep=True)
df["pmd_nr_id"] = df["pmd_id"]+df["nr"].apply(lambda x: "_"+str(int(x)))
sequences_df = df[['pmd_nr_id', 'crossref', 'seq']].drop_duplicates(keep="first")
print("#-of sequences b4 seq-len filter", sequences_df.shape[0])
protid_seq_tuple_list = [(tuple.pmd_nr_id, tuple.seq) for tuple in sequences_df.itertuples() if len(str(tuple.seq))<=1022]
print("#-of sequences after seq-len filter", len(protid_seq_tuple_list))

new_protein_acc_list = list(zip(*protid_seq_tuple_list))[0]
df = df[df["pmd_nr_id"].isin(new_protein_acc_list)]
print("after filtering on seq-len <=1022:", df.shape)
print(df["functional_effect"].value_counts())

data_filered_on_seqlen = df.copy(deep=True)
data_filered_on_seqlen.columns

df = data_filered_on_seqlen.copy(deep=True)
df.to_csv(home_dir+"models/aa_common/datasets_pmd_analysis/pmd.tsv", sep="\t", index=False)
sequences_df = df[['pmd_nr_id', 'crossref', 'seq']].drop_duplicates(keep="first")
print("#-of sequences", sequences_df.shape)

out_fasta_filepath = home_dir+"models/aa_common/datasets_pmd_analysis/pmd_sequences.fasta"
out_fasta_file_handle = open(out_fasta_filepath, "w")

for tuple in sequences_df.itertuples():
    out_fasta_file_handle.write(f">{tuple.pmd_nr_id} | {tuple.crossref}\n")
    out_fasta_file_handle.write(f"{tuple.seq}\n")
    # break
out_fasta_file_handle.close()  

### tests

# 'mut_id', 'md5', 'pmd_id', 'nr', 'mut_real', 'chrom', 'chrom_pos', 'ref_allele', 'alt_allele', 'function', 'source', 'crossref', 'snp_id'
print("#-of protein variants: ", df[['mut_id', 'md5', 'pmd_id', 'nr', 'mut_real']].drop_duplicates(keep="first").shape[0])
print("#-of chromosomal variants: ", df[['chrom', 'chrom_pos', 'ref_allele', 'alt_allele']].drop_duplicates(keep="first").shape[0])
print("#-of protein sequences: ", df[['pmd_id', 'nr']].drop_duplicates(keep="first").shape[0])
print("#-of rows not mapped to snp-ids: ", df[pd.isna(df["snp_id"])].shape[0])
print("#-of rows mapped to snp-ids: ", df[~pd.isna(df["snp_id"])].shape[0])
print("total: ", df.shape[0])
df["functional_effect"].value_counts()


z = data_filered_on_seqlen
for col in z.columns:
    print(col, z[pd.isna(z[col])].shape)
